add float

+++ monger/model/openai/request_body.go
@@ -6,5 +6,5 @@ type RequestBody struct {
 	Stream      bool      `json:"stream"`
 	Messages    []Message `json:"messages"`
 	MaxTokens   int       `json:"max_tokens"`
-	Temperature int       `json:"temperature"`
+	Temperature float32   `json:"temperature"`
 }
